Rename littlesmallmethods.js to more familiar prelude.js. Moved src/js/development/* to src/js/production/ because the development build passed all the tests.
