+++ src/js/production/expression.js
@@ -5,11 +5,22 @@
   optimization = require('./optimization');
   var Expression;
   Expression = function () {
+    this.bundledMethods = null;
     if (this.init) {
       this.init.apply(this, arguments);
     }
   };
   Expression.prototype = new LittleParser();;
+  Expression.prototype.init = function () {
+    var _this = this;
+    return _this.bundledMethods = [];
+  };
+  Expression.prototype.bundleAMethodIfAvailable = function (methodName) {
+    var _this = this;
+    return ((_this.bundledMethods.indexOf(methodName) > -1) && bundlableMethods.bundlable(methodName)) ? (function () {
+      return _this.bundledMethods.push(bundlableMethods.bundle(methodName));
+    })() : void 0;
+  };
   Expression.prototype.expression = function () {
     var _this = this;
     var tmpl;
@@ -20,10 +31,22 @@
         return _this.assignments();
       });
       cascade = _this.cascade();
-      return _this.templateapply(tmpl, {
-        "assignments": assignments,
-        "cascade": cascade
-      });
+      return (function () {
+        var _ret;
+        try {
+          _ret = (function () {
+            return _this.templateapply(tmpl, {
+              "assignments": assignments,
+              "cascade": cascade
+            });
+          })();
+        } catch (err) {
+          _ret = function (e) {
+            return console.log(e);
+          }(err);
+        }
+        return _ret;
+      })();
     });
   };
   Expression.prototype.assignments = function () {
@@ -51,7 +74,6 @@
         _this.notFollowedBy(function () {
           return _this.semicolon();
         });
-
         return se;
       }, function () {
         var conti;
